Update Seis.read_mseed for new Seis version
Update `parse_data_response` and the tests to call
`Seis.read_mseed` with the new Seis.jl calling convention of
putting the return type as the last argument.

This change will work on Seis.jl commits later than
810db13daaa144b3ae641c8e442c218121147112
but compatibility in this package should be updated once
v0.4 of Seis.jl is released.